beenews-api/requirements.txt
Flask==0.10.1
flask-cors==2.0.1
PyYAML>=3.11
uWSGI==2.0.10
unidecode==0.4.17
mongoengine==0.9.0
pymongo==2.8.1
